requirements.txt
streamlit>=1.29.0
pandas==2.1.3
numpy==1.26.2
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.3
beautifulsoup4==4.12.2
schedule==1.2.1
nbformat>=4.2.0
altair==5.1.2
nba_api==1.4.1
scikit-learn==1.3.2
